+++ src/utils/ai.tsx
@@ -3,7 +3,6 @@ import Typography from "@mui/material/Typography";
 
 import { emailConfig } from "../constants";
 import { AddAlert, Recipe, ReportErrorValues } from "../types";
-import { makeHeaders } from "./utility";
 
 export const generateRecipe = async (
   params: { [key: string]: string | number | boolean },
@@ -15,21 +14,29 @@ export const generateRecipe = async (
     .map(([key, value]) => `${key}=${encodeURIComponent(value)}`)
     .join("&");
 
-  try {
-    const headers = await makeHeaders(user);
+  const appCheckTokenResponse =
+    await user.auth.appCheckServiceProvider.instances
+      .get("[DEFAULT]")
+      .getToken(false)
+      .catch(onFailure);
 
-    fetch(`/api/generate-recipe?${searchParamsText}`, { headers })
-      .then((response) => {
-        if (!response.ok) {
-          throw Error(response.statusText);
-        }
+  const authorization = btoa(`${user.uid}:${user.accessToken}`);
 
-        return response.text();
-      })
-      .then(onSuccess);
-  } catch (error) {
-    onFailure(error);
-  }
+  fetch(`/api/generate-recipe?${searchParamsText}`, {
+    headers: {
+      Authorization: authorization,
+      "X-Firebase-AppCheck": appCheckTokenResponse.token,
+    },
+  })
+    .then((response) => {
+      if (!response.ok) {
+        throw Error(response.statusText);
+      }
+
+      return response.text();
+    })
+    .then(onSuccess)
+    .catch(onFailure);
 };
 
 export const parseResponse = (textResponse: string) => {
